+++ CHANGES.md
@@ -460,6 +460,10 @@ OpenSSL Releases
 
    *Igor Ustinov*
 
+ * Add IKEV2 KDF (EVP_KDF-IKEV2KDF) to EVP_KDF
+
+   *Helen Zhang*
+
 OpenSSL 3.6
 -----------